+++ src/mask.jl
@@ -13,7 +13,7 @@ function _Mask{1}(bits::Integer...)
     chunk = UInt64(0)
     for b in bits
         @check 1 ≤ b ≤ 64
-        offset = b - 1
+        offset = (b - 1) % UInt64
         chunk |= (UInt64(1) << offset)
     end
     return _Mask((chunk,))
@@ -24,7 +24,7 @@ function _Mask{M}(bits::T...) where {M,T<:Integer}
     for b in bits
         @check 1 ≤ b ≤ M * 64
         chunk = (b - 1) >>> 6
-        offset = (b - 1) & T(0x3F)
+        offset = ((b - 1) & T(0x3F)) % UInt64
         chunks = Base.setindex(chunks, chunks[chunk+1] | (UInt64(1) << offset), chunk + 1)
     end
     return _Mask(chunks)
@@ -38,7 +38,7 @@ function _Mask{1}(::_Not, bits::Integer...)
     chunk = typemax(UInt64)
     for b in bits
         @check 1 ≤ b ≤ 64
-        offset = b - 1
+        offset = (b - 1) % UInt64
         chunk &= ~(UInt64(1) << offset)
     end
     return _Mask((chunk,))
@@ -49,7 +49,7 @@ function _Mask{M}(::_Not, bits::T...) where {M,T<:Integer}
     for b in bits
         @check 1 ≤ b ≤ M * 64
         chunk = (b - 1) >>> 6
-        offset = (b - 1) & T(0x3F)
+        offset = ((b - 1) & T(0x3F)) % UInt64
         mask = ~(UInt64(1) << offset)
         chunks = Base.setindex(chunks, chunks[chunk+1] & mask, chunk + 1)
     end
@@ -107,15 +107,15 @@ end
 
 _is_not_zero(m::_Mask)::Bool = !_is_zero(m)
 
-function _active_bit_indices(mask::_Mask{M})::Vector{UInt8} where M
-    indices = UInt8[]
+function _active_bit_indices(mask::_Mask{M})::Vector{Int} where M
+    indices = Int[]
     for chunk_index in 1:M
         chunk = mask.bits[chunk_index]
-        base = UInt8((chunk_index - 1) * 64)
+        base = (chunk_index - 1) * 64
         while chunk != 0
             tz = trailing_zeros(chunk)
-            push!(indices, base + UInt8(tz + 1))
-            chunk &= chunk - 1  # clear lowest set bit
+            push!(indices, base + tz + 1)
+            chunk &= chunk - UInt64(1) # clear lowest set bit
         end
     end
     return indices
@@ -150,36 +150,36 @@ function _Mask(mask::_MutableMask)
     return _Mask(Tuple(mask.bits))
 end
 
-@inline function _set_bit!(mask::_MutableMask{1}, i::UInt8)
-    offset = i - UInt8(1)
-    val = UInt64(1) << (offset % UInt64)
+@inline function _set_bit!(mask::_MutableMask{1}, i::Int)
+    offset = (i - 1) % UInt64
+    val = UInt64(1) << offset
     mask.bits[1] |= val
 end
-@inline function _set_bit!(mask::_MutableMask, i::UInt8)
-    chunk = (i - UInt8(1)) >>> 6
-    offset = (i - UInt8(1)) & 0x3F
-    val = UInt64(1) << (offset % UInt64)
+@inline function _set_bit!(mask::_MutableMask, i::Int)
+    chunk = (i - 1) >>> 6
+    offset = ((i - 1) % UInt64) & UInt64(0x3F)
+    val = UInt64(1) << offset
     mask.bits[chunk+1] |= val
 end
 
-@inline function _clear_bit!(mask::_MutableMask{1}, i::UInt8)
-    offset = i - UInt8(1)
-    val = ~(UInt64(1) << (offset % UInt64))
+@inline function _clear_bit!(mask::_MutableMask{1}, i::Int)
+    offset = (i - 1) % UInt64
+    val = ~(UInt64(1) << offset)
     mask.bits[1] &= val
 end
-@inline function _clear_bit!(mask::_MutableMask, i::UInt8)
-    chunk = (i - UInt8(1)) >>> 6
-    offset = (i - UInt8(1)) & 0x3F
-    val = ~(UInt64(1) << (offset % UInt64))
+@inline function _clear_bit!(mask::_MutableMask, i::Int)
+    chunk = (i - 1) >>> 6
+    offset = ((i - 1) % UInt64) & UInt64(0x3F)
+    val = ~(UInt64(1) << offset)
     mask.bits[chunk+1] &= val
 end
 
-@inline function _get_bit(mask::Union{_Mask{1},_MutableMask{1}}, i::UInt8)::Bool
-    offset = i - UInt8(1) # which bit within that UInt64
-    return (mask.bits[1] >> (offset % UInt64)) & UInt64(1) == 1
+@inline function _get_bit(mask::Union{_Mask{1},_MutableMask{1}}, i::Int)::Bool
+    offset = (i - 1) % UInt64 # which bit within that UInt64
+    return ((mask.bits[1] >> offset) & UInt64(1)) % Bool
 end
-@inline function _get_bit(mask::Union{_Mask,_MutableMask}, i::UInt8)::Bool
-    chunk = (i - UInt8(1)) >>> 6 # which UInt64 (0-based)
-    offset = (i - UInt8(1)) & 0x3F # which bit within that UInt64
-    return (mask.bits[chunk+1] >> (offset % UInt64)) & UInt64(1) == 1
+@inline function _get_bit(mask::Union{_Mask,_MutableMask}, i::Int)::Bool
+    chunk = (i - 1) >>> 6 # which UInt64 (0-based)
+    offset = ((i - 1) % UInt64) & UInt64(0x3F) # which bit within that UInt64
+    return ((mask.bits[chunk+1] >> offset) & UInt64(1)) % Bool
 end